[REVERT ME] Disable Firefox testing
Since Firefox 63, geckodriver >= 0.21 is needed but
selenium-standalone uses 0.20.1 currently. [1]

selenium-standalone has updated the default conf [2] but not
released a new version yet.

Hopefully the dependency tree is setup so the new version will
be used as soon as it's released, or we have to override default
configs or wait for polymer-cli to update as well.

This commit can be reverted to enable Firefox testing once the
above is resolved.

[1]: https://github.com/mozilla/geckodriver/issues/1406
[2]: https://github.com/vvo/selenium-standalone/commit/455ea28586cff04fab0cb46af08e057fd401cbda

Signed-off-by: Patrik Kullman <patrik.kullman@neovici.se>