feat(haunted-polymer): use setProperties instead of outputPath
Setting a single outputPath is very limiting, and causes unnecessary recalculations in polymer when observing subpaths of the outputPath. For example `computed: '_computeFrom(state.invoices)'` results in a new value even if state.invoices did not change, but state.otherProp did.

Migration instructions:

Drop the outputPath from the hauntedPolymer call and move it to the output of the hook:

```js
hook = () => return 'a';

hauntedPolymer('state', hook)
```

becomes

```js
hook = () => return {state: 'a'};

hauntedPolymer(hook)
```