feat: better invoke
After reading up more on type inferrence, I realized that invoke was
typed badly. We can 100% rely on typescript doing type inference without
the need for conditional types. The missing piece was the type assertion
inside of the `is function` branch. The old version had a problem in
that it narrowed types too much. This version works exactly as you'd
expect it, including enforcing the types of the function arguments.